Optional NAT
This feature allows users to disable NAT gateway creation,
typically for cost-savings or fully-private networks.

I have tested this by hand in with NAT enabled and disabled by
creating an instance in the DMZ, shelling to it, reaching the
internet, then creating an instance in the admin layer, shelling
to it, and validating that it can (nat_enabled=true) or cannot
(nat_enabled=false) reach the internet.